test(ui): TE-1481 check if TE_DEV_PROXY_SERVER is not set

+++ thirdeye-ui/src/test/e2e/support/reset-datasource.command.js
@@ -13,16 +13,21 @@
  * the License.
  */
 Cypress.Commands.add("resetDatasets", () => {
+    if (process.env.TE_DEV_PROXY_SERVER !== undefined) {
+        throw new Error(
+            "TE_DEV_PROXY_SERVER is set. Failing in case it is linked to a dev server"
+        );
+    }
     cy.request({
         method: "DELETE",
-        url: "http://localhost:7004/api/data-sources/all",
+        url: "http://localhost:7005/api/data-sources/all",
     });
     cy.request({
         method: "DELETE",
-        url: "http://localhost:7004/api/datasets/all",
+        url: "http://localhost:7005/api/datasets/all",
     });
     cy.request({
         method: "DELETE",
-        url: "http://localhost:7004/api/metrics/all",
+        url: "http://localhost:7005/api/metrics/all",
     });
 });
